+++ amazingmovies-api/src/main/java/br/uece/eesdevops/amazingmovies/repository/EvaluationRepository.java
@@ -9,10 +9,10 @@ import org.springframework.stereotype.Repository;
 import br.uece.eesdevops.amazingmovies.domain.entity.Evaluation;
 
 @Repository
-public interface EvaluationRepository extends JpaRepository<Evaluation, Long>{
+public interface EvaluationRepository extends JpaRepository<Evaluation, Integer>{
 
 	@Modifying
 	@Query("delete from Evaluation e where e.movie.id=:movieId")
-	public void deleteByMovieId(@Param("movieId") Long movieId);
+	public void deleteByMovieId(@Param("movieId") Integer movieId);
 	
 }
